springpad-doc-parser/requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Environment Variable Management
python-dotenv
beautifulsoup4
requests